requirements.txt
flask
flask_login
psycopg2-binary
flask_mail
flask_bcrypt
psycogreen
gevent
ujson
requests
IPy
pygal
geoip2
Flask-WTF
python-dateutil
email_validator
connexion
WTForms==2.3.3
Flask-Reuploaded
Flask-WeasyPrint
iptools
